requirements.txt
torch
torchvision
facenet-pytorch
opencv-python
numpy
pillow
tqdm
mediapipe==0.10.9